Update $widthFactor in getBarcode for float
changing the field to float it is possible to get smaller sizes, using 0.5 for example

+++ src/BarcodeGeneratorSVG.php
@@ -15,7 +15,7 @@ class BarcodeGeneratorSVG extends BarcodeGenerator
      * @return string SVG code.
      * @public
      */
-    public function getBarcode($barcode, $type, int $widthFactor = 2, int $height = 30, string $foregroundColor = 'black')
+    public function getBarcode($barcode, $type, float $widthFactor = 2, int $height = 30, string $foregroundColor = 'black')
     {
         $barcodeData = $this->getBarcodeData($barcode, $type);
